requirements.txt
altgraph==0.17
APScheduler==3.6.3
astroid==2.5.2
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
comtypes==1.1.9
cycler==0.10.0
docxtpl==0.11.4
et-xmlfile==1.0.1
Flask==2.0.2
fpdf==1.7.2
future==0.18.2
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.28.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gTTS==2.2.2
httplib2==0.19.1
idna==2.10
isort==5.8.0
itsdangerous==2.0.1
jaraco.context==4.0.0
Jinja2==3.0.3
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.1
mccabe==0.6.1
more-itertools==8.7.0
numpy==1.20.2
oauthlib==3.1.0
openpyxl==3.0.7
packaging==20.9
pandas==1.2.3
pdfkit==0.6.1
pefile==2019.4.18
Pillow==8.2.0
playsound==1.2.2
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/admin/Downloads/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
pygame==2.0.1
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
pylint==2.7.4
pyparsing==2.4.7
pypiwin32==223
PySimpleGUI==4.39.1
pyTelegramBotAPI==4.3.1
python-dateutil==2.8.1
python-docx==0.8.10
python-telegram-bot==13.10
pyttsx3==2.90
pytube==10.6.1
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pywin32==300
pywin32-ctypes==0.2.0
reportlab==3.5.67
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
toml==0.10.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==2.0.2
wikipedia==1.4.0
wolframalpha==5.0.0
wrapt==1.12.1
xlrd==1.2.0
xmltodict==0.12.0
